miniAOD monitoring (this version was running in local tests since a while)

+++ comparisons/validateJR.sh
@@ -6,37 +6,61 @@ inList=$4
 cWD=`pwd`
 export pidList=""
 nProc=$(getconf _NPROCESSORS_ONLN)
+
+function waitForProcesses {  
+    pidList=${pidList}" "${!}
+    export pidList
+    echo $pidList
+    nRunning=`ps -p $pidList | grep -c "^[ ]*[1-9]"`
+    while [ "$nRunning" -ge "$nProc"  ]; do
+        nRunning=`ps -p $pidList | grep -c "^[ ]*[1-9]"`
+	echo "limiting number of parallel processes"
+        sleep 10
+    done
+    cd ${cWD}
+    echo $pidList > lastlist.txt
+}
+
+
 echo Start processing at `date`
 touch missing_map.txt
 ls -d ${baseA}/[1-9]* | sed 's|.*/||' | while read -r d; do
   grep  " $d/" ${inList} >& /dev/null || echo $d >> missing_map.txt
 done
 grep root ${inList} | grep -v "#" | while read -r dsN fN procN comm; do 
-    [ ! -f "${baseA}/${fN}" ] && echo Missing ${baseA}/${fN} && continue
-    extN=all_${diffN}_${dsN}
-    mkdir -p ${extN}
-    cd ${cWD}/${extN}
-    cp ~/tools/validate.C ./
-    echo "Will run on ${fN} in ${cWD}/${extN}"
-    echo "Now in `pwd`"
-    g++ -shared -o validate.so validate.C `root-config --cflags ` -fPIC
-    echo -e "gSystem->Load(\"libFWCoreFWLite.so\");\n AutoLibraryLoader::enable();\n 
+    [ ! -f "${baseA}/${fN}" ] && echo Missing ${baseA}/${fN}
+    #process regular files first
+    if [ -f "${baseA}/${fN}" ]; then
+	extN=all_${diffN}_${dsN}
+	mkdir -p ${extN}
+	cd ${cWD}/${extN}
+	cp ~/tools/validate.C ./
+	echo "Will run on ${fN} in ${cWD}/${extN}"
+	echo "Now in `pwd`"
+	g++ -shared -o validate.so validate.C `root-config --cflags ` -fPIC
+	echo -e "gSystem->Load(\"libFWCoreFWLite.so\");\n AutoLibraryLoader::enable();\n FWLiteEnabler::enable();\n 
     .x validate.C+(\"${extN}\", \"${baseA}/${fN}\", \"${baseB}/${fN}\", \"${procN}\");\n .qqqqqq" | root -l -b >& ${extN}.log &
-# manually set the make flags, not needed in most cases
-#    gSystem->SetMakeSharedLib(\"cd \$BuildDir ;g++  -c \$Opt -pipe -m64 -Wshadow -Wall -W -Woverloaded-virtual -fPIC -std=c++11 -Wno-deprecated-declarations -DG__MAXSTRUCT=36000 -DG__MAXTYPEDEF=36000 -DG__LONGLINE=4096 -pthread \$IncludePath \$SourceFiles ; g++ \$ObjectFiles -shared -Wl,-soname,\$LibName.so -m64 -Wl,--hash-style=gnu -O2  \$LinkedLibs -o \$SharedLib\");\n 
-#    cout<< gSystem->GetMakeSharedLib()<<endl ;\n
+	waitForProcesses
+    fi
+    #process miniAOD files now
+    fNBase=`echo ${fN} | sed -e 's/.root$//g'`
+    mFN="${fNBase}_inMINIAOD.root"
+    if [ ! -f "${baseA}/${mFN}" ]; then
+        mFN="${fNBase}_inMINIAODSIM.root"
+    fi
+    if [ -f "${baseA}/${mFN}" ]; then
+	echo $mFN
+	extmN=all_mini_${diffN}_${dsN}
+	mkdir -p ${extmN}
+	cd ${cWD}/${extmN}
+	cp ~/tools/validate.C ./
+	echo "Will run on ${mFN} in ${cWD}/${extmN}"
+	echo "Now in `pwd`"
+	echo -e "gSystem->Load(\"libFWCoreFWLite.so\");\n AutoLibraryLoader::enable();\n FWLiteEnabler::enable();\n
+        .x validate.C+(\"${extmN}\", \"${baseA}/${mFN}\", \"${baseB}/${mFN}\", \"${procN}\");\n .qqqqqq" | root -l -b >& ${extmN}.log &
+	waitForProcesses
+    fi
 
-    pidList=${pidList}" "${!}
-    export pidList
-    echo $pidList
-    nRunning=`ps -p $pidList | grep -c root`
-    while  [ "$nRunning" -ge "$nProc"  ]; do  
-	nRunning=`ps -p $pidList | grep -c root`
-        echo "limiting number of parallel processes"
-	sleep 10
-    done
-    cd ${cWD}
-    echo $pidList > lastlist.txt
 done
 allPids=`cat lastlist.txt`
 nRunning=1
